+++ h2gis-utilities/src/main/java/org/h2gis/utilities/JDBCUtilities.java
@@ -654,13 +654,12 @@ public class JDBCUtilities {
         final Statement statement = connection.createStatement();
         List<String> fieldValues = new ArrayList<>();
         try {
-            ResultSet result = statement.executeQuery("SELECT DISTINCT " + TableLocation.quoteIdentifier(fieldName) + " FROM " + TableLocation.parse(tableName).toString(dbType));
-            try {
+            try (ResultSet result = statement.executeQuery("SELECT DISTINCT " +
+                    TableLocation.quoteIdentifier(fieldName, dbType) + " FROM " +
+                    TableLocation.parse(tableName, dbType))) {
                 while (result.next()) {
                     fieldValues.add(result.getString(1));
                 }
-            } finally {
-                result.close();
             }
         } finally {
             statement.close();